rename and change this target framework to net8.0-windows
