+++ GakumasResize.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31613.86
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "StarlightResize", "StarlightResize\StarlightResize.csproj", "{5D9484BB-B605-4A82-A925-967751D86332}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GakumasResize", "GakumasResize\GakumasResize.csproj", "{5D9484BB-B605-4A82-A925-967751D86332}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
