+++ GakumasResize/Form1.resx
