+++ GakumasResize/NativeMethods.txt
