+++ GakumasResize/Program.cs
@@ -1,10 +1,7 @@
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using System.Windows.Forms;
 
-namespace StarlightResize
+namespace GakumasResize
 {
     static class Program
     {
